Fix parsing with more sophisticated pydantic model

+++ server/util/parse.py
@@ -1,13 +1,13 @@
 import re
 
 
-def extract_temperature(instruction: str) -> float | None:
+def extract_temperature(instruction: str) -> int | None:
     """Extract the temperature from a string."""
     print(f'Extracting temperature from instruction: {instruction}')
     match = re.search(r'\b(\d{1,3})\s*°?\s*(C|F)\b', instruction)
     if not match:
         return None
-    temperature = float(match.group(1))
+    temperature = int(match.group(1))
     unit = match.group(2)
     if unit == 'F':
         temperature = (temperature - 32) * 5 / 9
